Move run command to separate module
